fix: ignore empty csf v2 storyName
Prior to Storybook 6.4, a story with an empty string as a CSF v2-style
hoisted `storyName` and a (non-empty) CSF v1-style `story.name` would
have a display name generated from the export name, as if there were no
given story name at all. Starting with 6.4, the behavior has changed so
that the v2-style empty string name is disregarded and the v1-style name
is used instead.

Although this change would technically result in a different display
name between this extension and older versions of Storybook, this
scenario is fairly contrived and unlikely to appear in the real world.
So it doesn't appear to be worth adding any kind of version detection or
fallback behavior to replicate Storybook's exact behavior in this case.