fix: improve messaging when no stories found
Always show the stories view, even if a configuration cannot be
detected. This makes it clearer that the extension has loaded and
provides an opportunity to provide a hint to the user about what may be
wrong and how to resolve it.

Avoid suggesting that there was a problem parsing a Storybook
configuration unless an actual problem was encountered; previously, this
message could appear while scanning for stories, even if there was no
problem loading the configuration.