fix: improve start-storybook script detection
Searching the workspace via `workspace.findFiles` is subject to the
user's `search.followSymlinks` setting, which will cause the search to
fail on platforms that use symlinks in the `.bin` directory.

Instead of searching the entire workspace, use node module resolution to
check specific locations based on the location of the Storybook
configuration directory, if available. If this check fails or the
configuration directory itself cannot be located, then fall back to the
`findFiles`-based search.

Fixes #460