feat: support Storybook 7 autodocs
In Storybook 7, docs are now rendered as peers to stories in the left
navigation instead of appearing in a separate "docs" tab (as opposed to
"canvas"). In addition, these autodoc entries are only created depending
on the project (and/or component) configuration.

To accommodate this new pattern, update the stories view to match the
new hierarchy. This means parent tree nodes are no longer generally used
to preview docs; the new child entries are used for this purpose
instead.

The new behavior is only effective if the current project uses Storybook
7 or greater; projects using Storybook 6.x retain the old behavior.

Fixes #759