fix: always determine config file location
The location of the Storybook config always needs to be known in order
to enable certain commands, like opening the configuration file or
attempting to start a Storybook server. To ensure these commands are
properly enabled, always subscribe to the configuration location.

This addresses an issue where the configuration location would never be
determined if all of the settings (i.e., stories globs) ordinarily
dependent on the configuration location were overridden via VS Code
settings, thus causing those commands to never become enabled.