feat: support additional launch strategies
Add support for new launch strategies for launching a Storybook
developer server. This allows the server to be launched via direct
invocation (`storybook` for Storybook 7 or `start-storybook` for
Storybook 6), npm script, VS Code task, or custom script.

The existing settings for configuring the `start-storybook` binary path
and arguments are deprecated in favor of new options.

Fixes #571
Fixes #605