refactor: add storybook version detection
Add internal support for detecting the installed version of Storybook.

This is based on:
- a hidden editor setting
- the detected version of `@storybook/cli` or `storybook` installed in
  the project
- the detected version of `@storybook/cli` or `storybook` identified in
  a supported lockfile
- the detected version (or range) of `@storybook/cli` or `storybook` in
  `package.json`

If none of these approaches yield a version, a default version (~the
most recent stable release) is used as a fallback. This is expected to
be rare, since the heuristics above capture a range of scenarios
(including those where a project's dependencies haven't been installed
locally).

Version detection allows the introduction of breaking behavioral changes
without dropping support for older versions or requiring users to
manually specify the version they use (or its corresponding behaviors).