fix: avoid confusing outgoing ports with open ports
Sometimes, ports used as the destination for outgoing connections were
erroneously being treated as locally open ports, causing connecting to
the Storybook dev server to fail.

This could be readily reproduced by trying to launch Storybook without
CI mode enabled and with the requested port already in use, in which
case no local port was available, but an outbound request from a watched
process caused port 443 to be returned incorrectly.