fix: add workaround for storybook CLI not honoring environment variables
To prevent launching the Storybook dev server from opening a browser
window or (in the event of a port conflict) asking the user for
interactive input, the server is generally launched in CI mode, which
suppresses these behaviors. To ensure that Storybook detects the same
configuration file used by Story Explorer, its location is also passed
explicitly. Due to a bug in Storybook 7, these parameters cannot be set
via environment variables.

As a workaround, if an affected version of Storybook is detected and the
`storybook` launch strategy is in use, automatically inject the
corresponding CLI flags, i.e., `--ci` and '--config-dir', if they are
not already included. (Other launch strategies may require different
workarounds, so this behavior may not be appropriate in those cases.)