feat: support parsing TypeScript, ESM configs
Use esbuild-register to parse config files, which matches how Storybook
parses such files and allows for ESM (regardless of whether the
containing project itself otherwise uses ESM) and TypeScript, both of
which were otherwise not supported.

This approach uses the WASM build of esbuild, which isn't ideal from a
performance (or bundle size) perspective. However, it simplifies
packaging, and the performance drawbacks are not expected to be material
for this application. Future optimizations could provide for
platform-specific builds and/or a different approach to config file
parsing that could reduce the on-disk footprint.

Fixes #756